app/requirements.txt
Flask
Flask-CORS
python-dotenv
Werkzeug
langchain==0.1.0
openai==1.56.1
faiss-cpu==1.7.4
tiktoken==0.5.2